lj92: Fix assignment of SSSS for negative diff values
SSSS=16 should only be used for diff=-32678 as specified in H.1.2.2:
https://www.w3.org/Graphics/JPEG/itu-t81.pdf

This is a partial fix for the problems in the encoder and decoder for
SSSS=16 that shouldn't break anything. Just two improvements:
* Reduced size for the encoded buffer.
* Reduced risk of running into issues with decoders compliant with
the spec. (From now on SSSS=16 will not be used used for any
bitdepths lower than 15.)

Another problem is that diff values for SSSS=16 should be encoded
using zero bits. This would also be easy to fix (for both the encoder
and the decoder), but there needs to be a plan for how to handle
backwards compatibility.